+++ CHANGES.rst
@@ -10,6 +10,7 @@
   models. [#42]
 - Add converter for Spline1D model. [#43]
 - Add astropy Table connector for ASDF. [#47]
+- Move assert_model_equal to helpers module. [#50]
 
 0.1.2 (2021-12-14)
 ------------------